fix: unidades fraccionarias truncadas a 0 en reparto de lote (papel/tinta/pagos)
Hallazgo real al verificar el lote 66 recien finalizado: la orden 6869
(0.5 metros de DTF, producto medido por metro) quedo SIN ningun papel
ni tinta asignado -- intval($order['unidades_orden']) truncaba 0.5 a 0,
la proporcion de reparto le tocaba 0/total = 0%, y el consumo estimado
resultante (0) no pasaba el guard "if ($consumo_estimado > 0)".

Mismo bug de fondo ya corregido el 2026-08-31 en asignacion de personal
para esta misma orden (6869) -- nunca se replico en los 3 endpoints de
finalizacion de lote porque estaban huerfanos/sin uso real hasta hoy.
Se cambia intval->floatval en las 4 ocurrencias que afectan calculo
real (reparto proporcional de papel y tinta en finalizar-departamento,
finalizar-impresion y finalizar-corte). Se deja intval en pagos.cantidad
a proposito: esa columna es integer por esquema (no numeric), es solo
un campo de reporte/conteo, y monto_pago (el valor real del pago) ya se
calcula correctamente aparte via SQL con los valores exactos.

El stock total SI se descontaba correctamente (una sola resta sobre la
cantidad total del rollo, independiente del reparto) -- lo que faltaba
era la trazabilidad por orden en inventario_movimientos/tintas para las
ordenes con unidades fraccionarias.

+++ app/routes/manufacturing.php
@@ -1703,7 +1703,12 @@ return function (App $app) {
           if ($total_unidades_segmento > 0) {
             foreach ($ordenes_para_distribuir as $order) {
               $id_orden_actual = $order['id_orden'];
-              $unidades_orden = intval($order['unidades_orden']);
+              // floatval, no intval: hay productos medidos en unidades
+              // fraccionarias (ej. DTF por metros, 0.5) -- truncar a entero
+              // dejaba la orden con proporcion 0 y sin ningun consumo
+              // asignado (mismo bug ya corregido en asignación de personal
+              // el 2026-08-31 para la orden 6869, nunca replicado aquí).
+              $unidades_orden = floatval($order['unidades_orden']);
               $proporcion = $unidades_orden / $total_unidades_segmento;
               $consumo_estimado_orden = $cantidad_total_consumida * $proporcion;
 
@@ -2091,7 +2096,10 @@ return function (App $app) {
 
           if ($total_unidades_segmento > 0) {
             foreach ($ordenes_para_distribuir as $order) {
-              $proporcion = intval($order['unidades_orden']) / $total_unidades_segmento;
+              // floatval, no intval: ver comentario en /finalizar-departamento
+              // -- productos por metro (ej. DTF 0.5m) quedaban sin ningún
+              // papel asignado por truncarse a 0.
+              $proporcion = floatval($order['unidades_orden']) / $total_unidades_segmento;
               $consumo_estimado = $cantidad_total_papel * $proporcion;
 
               if ($consumo_estimado > 0) {
@@ -2109,7 +2117,8 @@ return function (App $app) {
         $colores = $tinta['colores'] ?? []; // Mapa de [id_color => cantidad]
 
         foreach ($ordenes_del_lote as $order) {
-          $proporcion = intval($order['unidades_orden']) / $gran_total_unidades_lote;
+          // floatval, no intval: ver comentario en /finalizar-departamento.
+          $proporcion = floatval($order['unidades_orden']) / $gran_total_unidades_lote;
           foreach ($colores as $id_color => $cantidad) {
             $cantidad_proporcional = floatval($cantidad) * $proporcion;
             if ($cantidad_proporcional > 0) {
@@ -2321,7 +2330,8 @@ return function (App $app) {
         if ($total_unidades_segmento > 0) {
           foreach ($ordenes_para_distribuir as $order) {
             $id_orden_actual = $order['id_orden'];
-            $unidades_orden = intval($order['unidades_orden']);
+            // floatval, no intval: ver comentario en /finalizar-departamento.
+            $unidades_orden = floatval($order['unidades_orden']);
             $proporcion = $unidades_orden / $total_unidades_segmento;
 
             // Distribuir y registrar consumo
